Work around a register allocation problem.
The register allocator fails to correctly spill/restore
variable set to pre-allocated EirValue register when this one is caller
saved. On SPARC, the only instance of that is the safepoint latch.
This case doesn't seem to occur on AMD64.

Simply making the pre-allocated register live across the entire
method (by defining a synthetic variable in the prologue and a use
in the epilogue) is insufficient. An additional problem was in the
definition of the EirOperand in EirCall. This failed to capture that
the one of the caller-saved register is a pre-allocated value. Fixing
this correctly caused the register allocator to produce the desired spill, but
break subsequent uses  by loading the spilled value into an arbitrary
allocatable register, leaving the safepoint latch register in an
clobbered state. The problem is once the pre-allocated value is
redefined in an EirCall as a caller saved register, it isn't
propagated anymore. Consequently, the variables (that resulted from
earlier splits) derived from the pre-allocated value aren't propagated
nor marked in any special way as requiring the pre-allocated
location. Marking the operand of instructions bound to these variables
with the safepoint latch register as prefered register doesn't make
any difference. The changes required for fixing this problem need to
go much deeper.  So I reverse all that back to the original code.

Instead,in order to progress with running Hello on SPARC,
I use the following work-around: a local register is reserved to save
the safepoint latch around native call when the compiled method issue
an native call.
This works as follow: an eir call that use an ABI that requires
the safepoint latch to be caller saved mark the EirMethodGeneration
object as such. Upon creating a register allocator object, this one
checked if a local need to be reserved for  saving the safepoint
latch, and if so, exclude the local register from  the allocatable
set. When emitting a CALL instruction, the emitter check if the call
needs to save the safepoint latch and if so, use the local register.