Removed need for UninitializedObject when running the IR interpreters that include object instantiation. Instead, Unsafe.allocateInstance() is used. All uses of the latter now go through Objects.java.
Improved toString() of threads in the TeleInterpreter to show the complete stack trace of the thread.

Grafted TeleInterpreter onto the IrInterpreter framework. As a result, the TeleInterpreter can now run all the standard bytecode JUnit tests. This revealed some bugs in the interpreter which have all been fixed. These tests are now part of the gate.

Fixed disassembly of the tableswitch and lookupswitch bytecode instructions in the BytecodePrinter.