Added tracing for waiting and notifying monitors (included in the trace generated by the -XX:TraceMonitors option).
Fixed bug in StandardJavaMonitor.monitorWait where the list of waiting threads was not being updated properly when the wait completed as a result of a time out (instead of a notification).

Added tracing for the processing of weak references (included in the trace generated by the -XX:TraceGC option).

Fixed bug in processing of weak references where the java.lang.ref.Reference.discovered fiedl was not being cleared appropriately.

Fixed the ability to pause the VM from the inspector on Darwin-AMD64.

Made the default debug mode in the inspector be asynchronous.

Modified 'inspect' command in bin/max to accept the -XX:LogFile psuedo VM option.

Fixed reference to bin/mod-macosx-javalib.sh in MaxineTester.

Added -verbose:props VM option to display the system properties initialized during startup.

Added the directory containing the maxvm executable to the value of the sun.boot.library.path system property.

Temporary fix for Thread.interrupt(). I think this whole system needs to be double-checked against HotSpot implementation.