Added optimization to remove redundant UnsafeCasts. A redundant unsafe cast is one whose input and output kind are equivalent at the machine code level in terms of size and refmap categorization.
Added field to C1XCompilation to record if the current compilation is type-safe. A compilation is type-safe if it has not encountered any input such as use of the UNSAFE_CAST extended bytecode that breaks type safety. Sanity checking of type safety is disabled in compilations marked type unsafe. Note that this checking is not the same as that performed by the bytecode verifier. It exists only to catch violations of the invariant that all code fed to the compiler must have been verified.

Fixed bug in block skipping optimization.

Added notion of a canonical architecture kind for every kind value. This allows the Word kind to be treated as architecture-equal to Int on 32-bit platforms and Long on 64-bit platforms.

Modified FieldActor so that C1X also implements the @CONSTANT_WHEN_NOT_ZERO annotation correctly.

Fixed dead-code elimination to actually remove HIR nodes from the graph.