Fixed issues that caused generation of too many nops when enforcing patchability of direct calls. 1. Extended CPS's assembler support for alignment directive so that one can now specify a directive to align only if space left in the current aligned block of memory is not large enough to fit the next instructions.
2. Change patching requirements for displacement. It seems enough to write any
4-byte displacement as long as they fit within a 8-byte aligned 8-byte word.

3. Changed how alignment are enforced for the template-based JIT to reduced the amount of nops issued because of
call patching alignment constraints. Basically, a template with just one direct patchable call site doesn't requires alignment
of the whole template if its one direct call instruction will fit in a cache line when the template is appended to the current
position in the code buffer.
Experiments with DaCapo's fop and luindex shows that this effectively trim the number of nops issued in JITed code:
 3.4 times less nops (resp. 3.8) for luindex (fop). These savings comes on top of the other savings resulting from 1 and 2 above.