Completed supported for removing array bounds checks as part of canonicalization. Note that this is not the same as ABCE achieved by hoisting checks out of loops.
Normalized mechanism in HIR nodes by which nodes are notified when a runtime check is eliminated. This now means that each elimination offers a chance for removal of state no longer needed (such as a frame state snapshot).

Fixed bugs due to null ClassMethodActor for TargetMethods that were not produced by a compiler. This includes stubs, trampolines and adapters.

Disabled bytecode generation for exception dispatching and synchronized methods if the CPS or template JIT are not part of the VM configuration.

Added mechanism (SourceFrameVisitor) for iterating over the sources frames (or vframes in HotSpot terminology) of a call stack. Replaced all other source frame walking usages with this mechanism.