Changed native method linking so that it is performed when a native stub is compiled/generated. The only exception to this are methods compiled at boot image time as linking isn't meaningful in this context.
Made the Java frame anchor in a native method stub record the exact position of the native function call. This removes the brittle heuristic in SatckFrameWalker for finding the debug info correlated with the native function call site.

Modified initialization of DynamicLinker to remove the recursion when resolving native linking functions.

Added JDK_java_lang_Shutdown.java to bypass the native call to JVM_Halt when exiting the VM.

Removed HeapScheme.runFinalization(). This is JDK functionality.

Added output test for Runtime.runFinalizersOnExit().